requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-app-settings==0.7.1
django-appconf==1.0.4
django-braces==1.14.0
django-embed-video==1.4.0
idna==2.10
pilkit==2.0
Pillow==8.1.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3

dj-database-url==0.5.0
django-heroku==0.3.1
gunicorn==20.0.4
python-decouple==3.3
import==0.0.1
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
